dbeaver/pro#9396 Fix export zero date and zero timestamp from db to db
